chore: move tracer module out of wasmhost
