+++ jolt-core/src/jolt_onnx/mod.rs
@@ -10,5 +10,6 @@ pub mod instruction;
 pub mod onnx_host;
 pub mod subtable;
 pub mod trace; // TODO: Remove this module
+pub mod tracer;
 pub mod utils;
 pub mod vm;
